global 类型问题

+++ lib/page/models/global_model.dart
@@ -1,5 +1,5 @@
 class GlobalInfo {
-  int? appBarHeight;
+  double? appBarHeight;
   String? baseUrl;
 
   GlobalInfo({this.appBarHeight, this.baseUrl});
